refactor(plate): Merge file into plate
