--- tools/file/docs/README.md
@@ -1,31 +1,0 @@
-# @dxos/file
-
-
-
-## Dependency Graph
-
-```mermaid
-%%{ init: {'flowchart':{'curve':'basis'}} }%%
-
-flowchart LR
-
-%% Classes
-classDef def fill:#fff,stroke:#333,stroke-width:1px
-classDef root fill:#fff,stroke:#333,stroke-width:4px
-
-%% Nodes
-
-subgraph undefined [undefined]
-  style undefined fill:#faf7eb,stroke:#333
-  dxos/file("@dxos/file"):::root
-  click dxos/file "dxos/dxos/tree/main/tools/file/docs"
-end
-
-%% Links
-linkStyle default stroke:#333,stroke-width:1px
-```
-
-## Dependencies
-
-| Module | Direct |
-|---|---|
