--- tools/file/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@dxos/file",
-  "version": "2.33.8",
-  "private": true,
-  "main": "dist/src/index.js",
-  "dependencies": {
-    "mkdirp": "^1.0.4"
-  },
-  "devDependencies": {
-    "@types/mkdirp": "^1.0.2",
-    "@types/node": "^16.11.58",
-    "typescript": "^4.7.2"
-  }
-}
