+++ tools/plate/package.json
@@ -4,16 +4,17 @@
   "private": "true",
   "main": "dist/src/index.js",
   "dependencies": {
-    "@dxos/file": "workspace:*",
     "esbuild": "~0.14.1",
     "lodash.flatten": "^4.4.0",
     "lodash.zip": "^4.2.0",
+    "mkdirp": "^1.0.4",
     "prettier": "^2.7.1",
     "recursive-readdir": "^2.2.2"
   },
   "devDependencies": {
     "@types/lodash.flatten": "^4.4.6",
     "@types/lodash.zip": "^4.2.6",
+    "@types/mkdirp": "^1.0.2",
     "@types/node": "^16.11.58",
     "@types/prettier": "^2.7.1",
     "@types/recursive-readdir": "^2.2.0",
