+++ tools/plate/src/file/File.ts
