+++ tools/plate/src/file/index.ts
